moved queries into the database file
